cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arches-toolkit"
version = "0.1.0"
description = "CLI for the Flax & Teal Arches container toolkit"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Flax & Teal", email = "info@flaxandteal.co.uk" },
]
keywords = ["arches", "docker", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "tomlkit>=0.13",
    "rich>=13.7",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
arches-toolkit = "arches_toolkit.main:app"

[project.urls]
Homepage = "https://github.com/flaxandteal/arches-toolkit"
Source = "https://github.com/flaxandteal/arches-toolkit"
Issues = "https://github.com/flaxandteal/arches-toolkit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/arches_toolkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
